pyproject.toml
[tool.pdm]

[project]
name = "tvmer"
version = "0.1.5dev3"
description = ""
authors = [
    {name = "zrr1999", email = "2742392377@qq.com"},
]
dependencies = [
    "typer",
    "rich"
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
tvm = [
    "apache-tvm"
]

[project.scripts]
tvmer = "tvmer.__main__:app"

[build-system]
requires = ["pdm-pep517>=1.0"]
build-backend = "pdm.pep517.api"
